app/src/main/java/com/example/robolectric/NewFragment.java
package com.example.robolectric;

import android.app.Fragment;
import android.os.Bundle;
import android.support.annotation.Nullable;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;

/**
 * Created by Anand on 08/10/2016.
 */

public class NewFragment extends Fragment {
    @Nullable
    @Override
    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
        View layout = inflater.inflate(R.layout.fragment_new, container);
        return layout;
    }

    public static NewFragment newInstance() {
        return new NewFragment();
    }
}
